# communication/fault-tolerance/fault-tolerance-book/src/main/java/org/agoncal/fascicle/quarkus/faulttolerance/book/BookResource.java
package org.agoncal.fascicle.quarkus.faulttolerance.book;

import net.datafaker.Faker;
// tag::adocCircuitBreaker[]
import org.eclipse.microprofile.faulttolerance.CircuitBreaker;

// end::adocCircuitBreaker[]
// tag::adocFallback[]
import org.eclipse.microprofile.faulttolerance.Fallback;

// end::adocFallback[]
// tag::adocTimeout[]
import org.eclipse.microprofile.faulttolerance.Timeout;

// end::adocTimeout[]
import org.eclipse.microprofile.rest.client.inject.RestClient;
import org.jboss.logging.Logger;

import jakarta.enterprise.context.ApplicationScoped;
import jakarta.json.Json;
import jakarta.json.JsonObject;
import jakarta.ws.rs.GET;
import jakarta.ws.rs.POST;
import jakarta.ws.rs.Path;
import jakarta.ws.rs.Produces;
import jakarta.ws.rs.core.MediaType;

import java.time.Instant;

// @formatter:off
// tag::adocSnippet[]
@Path("/books")
@Produces(MediaType.APPLICATION_JSON)
@ApplicationScoped
public class BookResource {

  // tag::adocSkip[]
  private static final Logger logger = Logger.getLogger(BookResource.class);

  private Faker faker = new Faker();
  // end::adocSkip[]
  @RestClient
  NumberProxy numberProxy;

  // tag::adocFallback[]
  @GET
  @Path("/numbers")
  @Fallback(fallbackMethod = "fallbackGenerateBookNumbers")
  public JsonObject generateBookNumbers() {

    // tag::adocSkip[]
    logger.info("Generating book numbers");
    // end::adocSkip[]
    // Invoking microservices
    IsbnNumber isbnNumber = numberProxy.generateIsbn(true);
    JsonObject issnNumber = numberProxy.generateIssn();

    return Json.createObjectBuilder()
      .add("isbn13", isbnNumber.isbn13)
      .add("gs1", isbnNumber.gs1)
      .add("isbn10", issnNumber.getJsonString("isbn10").getString())
      .build();
  }

  private JsonObject fallbackGenerateBookNumbers() {
    // tag::adocSkip[]
    logger.warn("Falling back on generating book numbers");
    // end::adocSkip[]
    return Json.createObjectBuilder()
      .add("isbn13", "dummy isbn")
      .add("gs1", "dummy gs1")
      .add("isbn10", "dummy issn")
      .build();
  }
  // end::adocFallback[]
  private JsonObject anotherFallbackMethod() {
    logger.warn("anotherFallbackMethod");
    return Json.createObjectBuilder()
      .add("isbn13", "anotherFallbackMethod")
      .add("gs1", "anotherFallbackMethod")
      .add("isbn10", "anotherFallbackMethod")
      .build();
  }
  // tag::adocTimeout[]
  @POST
  @Timeout(250)
  @Fallback(fallbackMethod = "fallbackCreateBook")
  public Book createBook() {
    // tag::adocSkip[]
    logger.info("Creating book");
    // end::adocSkip[]

    // Invoking microservice
    JsonObject issnNumber = numberProxy.generateIssn();

    Book book = new Book();
    book.title = faker.book().title();
    book.issn = issnNumber.getString("isbn10");
    book.generatedAt = Instant.now();

    return book;
  }

  private Book fallbackCreateBook() {
    // tag::adocSkip[]
    logger.warn("Falling back on creating a book");
    // end::adocSkip[]
    Book book = new Book();
    book.title = "dummy title";
    book.issn = "dummy issn";
    book.generatedAt = Instant.now();
    return book;
  }
  // end::adocTimeout[]
  // tag::adocSkip[]
  @Path("/legacy")
  @Timeout(250)
  // end::adocSkip[]
  // tag::adocCircuitBreaker[]
  @POST
  @Fallback(fallbackMethod = "fallbackCreateLegacyBook")
  @CircuitBreaker(requestVolumeThreshold = 4, failureRatio = 0.5,
                  delay = 2000, successThreshold = 2)
  public Book createLegacyBook() {
    // tag::adocSkip[]
    logger.info("Creating a legacy book");
    // end::adocSkip[]

    // Invoking microservice
    JsonObject issnNumber = numberProxy.generateIssn();

    Book book = new Book();
    book.title = faker.book().title();
    book.issn = issnNumber.getString("isbn10");
    book.generatedAt = Instant.now();

    return book;
  }

  private Book fallbackCreateLegacyBook() {
    // tag::adocSkip[]
    logger.warn("Falling back on creating a legacy book");
    // end::adocSkip[]
    Book book = new Book();
    book.title = "dummy legacy title";
    book.issn = "dummy legacy issn";
    book.generatedAt = Instant.now();
    return book;
  }
  // end::adocCircuitBreaker[]
}
// end::adocSnippet[]