# http/open-api/open-api-advanced/openapi.yaml
# tag::adocInfo[]
openapi: 3.1.0
info:
  title: Book API
  description: This API allows CRUD operations on books
  contact:
    name: '@agoncal'
    url: https://twitter.com/agoncal
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: "1.0"
externalDocs:
  description: All the Quarkus fascicle code
  url: https://github.com/agoncal/agoncal-fascicle-quarkus
# end::adocInfo[]
# tag::adocServer[]
servers:
  - url: "http://{host}.vintage-store/{port}"
    description: Vintage Store server 1
    variables:
      host:
        default: localhost
        description: Vintage Store main server
      port:
        default: "80"
        description: Vintage Store listening port
# end::adocServer[]
# tag::adocTag[]
tags:
  - name: api
    description: Public API that can be used by anybody
  - name: books
    description: Anybody interested in books
# end::adocTag[]
# tag::adocResponse[]
paths:
  /api/books:
    get:
      summary: Returns all the books from the database
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        "204":
          description: No books
    # end::adocResponse[]
    put:
      summary: Updates an exiting  book
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
        required: true
      responses:
        "200":
          description: The updated book
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
    # tag::adocRequestBody[]
    post:
      summary: Creates a book
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
        required: true
      # end::adocRequestBody[]
      responses:
        "201":
          description: The URI of the created book
          content:
            application/json:
              schema:
                format: uri
                type: string
  /api/books/ping:
    get:
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
  # tag::adocOperation[]
  /api/books/random:
    get:
      summary: Returns a random book
      description: Each time this API is invoked, a random book is returned from the database
      responses:
        "200":
          description: OK
  # end::adocOperation[]
  # tag::adocParameter[]
  /api/books/{id}:
    get:
      summary: Returns a book for a given identifier
      parameters:
        - name: id
          in: path
          description: Book identifier
          required: true
          schema:
            format: int64
            type: integer
      # end::adocParameter[]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        "404":
          description: The book is not found for the given identifier
    delete:
      summary: Deletes an exiting book
      parameters:
        - name: id
          in: path
          description: Book identifier
          required: true
          schema:
            format: int64
            type: integer
      responses:
        "204":
          description: The book has been successfully deleted
# tag::adocSchema[]
components:
  schemas:
    Book:
      description: Book representation
      required:
        - id
        - title
        - isbn
      type: object
      properties:
        id:
          format: int64
          type: integer
          readOnly: true
        title:
          type: string
        isbn:
          type: string
          example: "9798629562115"
        author:
          type: string
        price:
          type: number
        description:
          type: string
# end::adocSchema[]