# http/open-api/open-api-advanced/src/main/java/org/agoncal/fascicle/quarkus/http/openapi/advanced/Book.java
package org.agoncal.fascicle.quarkus.http.openapi.advanced;

import org.eclipse.microprofile.openapi.annotations.media.Schema;

import java.math.BigDecimal;

// tag::adocSchema[]
@Schema(name = "Book", description = "Book representation")
public class Book {

  @Schema(required = true, readOnly = true)
  public Long id;
  @Schema(required = true)
  public String title;
  @Schema(required = true, examples = "9798629562115")
  public String isbn;
  public String author;
  public BigDecimal price;
  public String description;
}
// end::adocSchema[]